Da int a double
Cambiato tipo di ritorno per il metodo che calcola l'area da int a double

+++ 22 marzo/AreaTriangolo/src/area_triangolo/AreaArray.java
@@ -2,9 +2,9 @@ package area_triangolo;
 
 public class AreaArray implements AreaInterface{
 
-	private int[] array = {5, 6};
+	private double[] array = {5, 6};
 	@Override
-	public int calcolaArea() {
+	public double calcolaArea() {
 		System.out.println("Area calcolata da array");
 		return (array[0] * array[1])/2;
