+++ 22 marzo/AreaTriangolo/src/area_triangolo/AreaInterface.java
@@ -2,5 +2,5 @@ package area_triangolo;
 
 public interface AreaInterface {
 
-	int calcolaArea();
+	double calcolaArea();
 }
